Update TF test names to follow standard
