+++ platform-tests/src/test/java/org/eclipse/deeplearning4j/frameworkimport/tensorflow/TestTFGraphAllSameDiff.java
@@ -1,40 +1,51 @@
 /*
- *
- *  *  ******************************************************************************
- *  *  *
- *  *  *
- *  *  * This program and the accompanying materials are made available under the
- *  *  * terms of the Apache License, Version 2.0 which is available at
- *  *  * https://www.apache.org/licenses/LICENSE-2.0.
- *  *  *
- *  *  *  See the NOTICE file distributed with this work for additional
- *  *  *  information regarding copyright ownership.
- *  *  * Unless required by applicable law or agreed to in writing, software
- *  *  * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- *  *  * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- *  *  * License for the specific language governing permissions and limitations
- *  *  * under the License.
- *  *  *
- *  *  * SPDX-License-Identifier: Apache-2.0
- *  *  *****************************************************************************
- *
- *
+ *  ******************************************************************************
+ *  *
+ *  *
+ *  * This program and the accompanying materials are made available under the
+ *  * terms of the Apache License, Version 2.0 which is available at
+ *  * https://www.apache.org/licenses/LICENSE-2.0.
+ *  *
+ *  *  See the NOTICE file distributed with this work for additional
+ *  *  information regarding copyright ownership.
+ *  * Unless required by applicable law or agreed to in writing, software
+ *  * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ *  * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ *  * License for the specific language governing permissions and limitations
+ *  * under the License.
+ *  *
+ *  * SPDX-License-Identifier: Apache-2.0
+ *  *****************************************************************************
  */
 
 package org.eclipse.deeplearning4j.frameworkimport.tensorflow;
 
 import lombok.extern.slf4j.Slf4j;
-import org.junit.jupiter.api.extension.ConditionEvaluationResult;
-import org.junit.jupiter.api.extension.ExecutionCondition;
-import org.junit.jupiter.api.extension.ExtensionContext;
+import lombok.val;
+import org.junit.jupiter.api.*;
 
-import java.util.Arrays;
-import java.util.List;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
+import org.nd4j.common.tests.tags.TagNames;
+import org.nd4j.linalg.api.ndarray.INDArray;
+import org.nd4j.linalg.factory.Nd4j;
+import org.nd4j.common.primitives.Pair;
+
+import java.io.File;
+import java.io.IOException;
+import java.util.*;
+import java.util.stream.Stream;
 
 import static org.junit.jupiter.api.Assumptions.assumeFalse;
 
 @Slf4j
-public class TFImportDisableModelsCondition implements ExecutionCondition {
+@Tag(TagNames.TENSORFLOW)
+public class TestTFGraphAllSameDiff {   //Note: Can't extend BaseNd4jTest here as we need no-arg constructor for parameterized tests
+
+    private static final TFGraphTestAllHelper.ExecuteWith EXECUTE_WITH = TFGraphTestAllHelper.ExecuteWith.SAMEDIFF;
+    private static final String BASE_DIR = "tf_graphs/examples";
+    private static final String MODEL_FILENAME = "frozen_model.pb";
 
     /**
      * NOTE: If this is empty or the tests names are wrong,
@@ -42,18 +53,20 @@ public class TFImportDisableModelsCondition implements ExecutionCondition {
      * the status of the test failing. No tests will run.
      */
     public final static List<String> EXECUTE_ONLY_MODELS = Arrays.asList(
-            /*"layers_dropout/rank2_d01_train",
-            "layers_dropout/rank4_d05_train",
-            "layers_dropout/rank3_d05_train_mask2",
-            "layers_dropout/rank4_d05_train_mask",
-            "layers_dropout/rank3_d05_train_mask1",
-            "layers_dropout/rank2_d09_train",
-            "layers_dropout/rank2_d05_train",*/
-
-
     );
 
     public static final String[] IGNORE_REGEXES = new String[]{
+            //crashes JVM
+            "scatter_nd_sub/unique_idxs/rank3shape_2indices",
+
+            //
+            //invalid graph:  Unable to run session input_0:0 is both fed and fetched.
+            //also due to the dynamic inputs being -1 3 for the first matrix multiply for the first 2 inputs
+            //the only valid batch size is 3.
+            "math_mul_order",
+            //points to a file with a URL, needs additional work
+            "compression_residual_gru",
+
             //Failing 2019/09/11 - https://github.com/eclipse/deeplearning4j/issues/7965
             // Still failing 2020/04/27 java.lang.IllegalStateException: Requested output variable Bincount does not exist in SameDiff instance
             //Invalid test cases. Verified by running graph against actual TF.
@@ -104,7 +117,7 @@ public class TFImportDisableModelsCondition implements ExecutionCondition {
             //08.05.2020 - https://github.com/eclipse/deeplearning4j/issues/8927
             "random_gamma/.*",
 
-            //08.05.2020 - https://github.com/eclipse/deeplearning4j/issues/8928
+            //08.05.2020 - https://github.cMatchCondom/eclipse/deeplearning4j/issues/8928
             "Conv3DBackpropInputV2/.*",
 
 
@@ -129,13 +142,23 @@ public class TFImportDisableModelsCondition implements ExecutionCondition {
 
 
 
-    private ExtensionContext.Store getStore(ExtensionContext context) {
-        return context.getStore(ExtensionContext.Namespace.create(getClass(), context.getRequiredTestMethod()));
+    public static Stream<Arguments> data() throws IOException {
+        val localPath = System.getenv(TFGraphTestAllHelper.resourceFolderVar);
+
+        // if this variable isn't set - we're using dl4j-tests-resources
+        if (localPath == null) {
+            File baseDir = new File(System.getProperty("java.io.tmpdir"), UUID.randomUUID().toString());
+            List<Object[]> params = TFGraphTestAllHelper.fetchTestParams(BASE_DIR, MODEL_FILENAME, EXECUTE_WITH, baseDir);
+            return params.stream().map(input -> Arguments.of(input));
+        } else {
+            File baseDir = new File(localPath);
+            return TFGraphTestAllHelper.fetchTestParams(BASE_DIR, MODEL_FILENAME, EXECUTE_WITH, baseDir).stream().map(input -> Arguments.of(input));
+        }
     }
 
-    @Override
-    public ConditionEvaluationResult evaluateExecutionCondition(ExtensionContext extensionContext) {
-        String modelName = getStore(extensionContext).toString();
+    @ParameterizedTest(name = "{2}")
+    @MethodSource("data")
+    public void testOutputOnly(Map<String, INDArray> inputs, Map<String, INDArray> predictions, String modelName, File localTestDir) throws Exception {
         if(EXECUTE_ONLY_MODELS.isEmpty()) {
             for(String s : IGNORE_REGEXES)  {
                 if(modelName.matches(s)) {
@@ -150,6 +173,33 @@ public class TFImportDisableModelsCondition implements ExecutionCondition {
         }
 
 
-        return ConditionEvaluationResult.disabled("Method found");
+
+        Pair<Double,Double> precisionOverride = TFGraphTestAllHelper.testPrecisionOverride(modelName);
+        Double maxRE = (precisionOverride == null ? null : precisionOverride.getFirst());
+        Double minAbs = (precisionOverride == null ? null : precisionOverride.getSecond());
+
+        boolean verboseDebugMode = true;
+        if(debugModeRegexes != null) {
+            for(String regex : debugModeRegexes) {
+                if(modelName.matches(regex)){
+                    verboseDebugMode = true;
+                    break;
+                }
+            }
+        }
+
+        try {
+            // TFGraphTestAllHelper.checkIntermediate(inputs,modelName,BASE_DIR,MODEL_FILENAME,EXECUTE_WITH,TFGraphTestAllHelper.LOADER,maxRE,minAbs,localTestDir,true);
+            Nd4j.getExecutioner().enableDebugMode(true);
+            Nd4j.getExecutioner().enableVerboseMode(true);
+            TFGraphTestAllHelper.checkOnlyOutput(inputs, predictions, modelName, BASE_DIR, MODEL_FILENAME, EXECUTE_WITH, new TFGraphTestAllHelper.DefaultGraphLoader(inputs), maxRE, minAbs, verboseDebugMode);
+        } catch (Throwable t){
+            log.error("ERROR Executing test: {} - input keys {}", modelName, (inputs == null ? null : inputs.keySet()), t);
+            throw t;
+        }
+        //TFGraphTestAllHelper.checkIntermediate(inputs, modelName, EXECUTE_WITH);
     }
+
+
+
 }
